requirements.txt
asgiref==3.8.1
attrs==23.2.0
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
django-crontab==0.7.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
